Changed http verb from post to put for adding to basket

+++ app/com/wehkart/controller/Messages.scala
@@ -5,7 +5,7 @@ object Messages {
 
   val ProductOutOfStock     = "Sorry, we're out of stock for this product"
   val ProductNotEnoughStock = "Sorry, we don't have enough stock for this product"
-  val ProductInvalidAmount  = (amount: Int) => s"You're trying to add an invalid number of products: $amount"
+  val ProductInvalidAmount  = (amount: Long) => s"You're trying to add an invalid number of products: $amount"
 
   val ProductNotInBasket = "You're trying to delete a product that is not your basket"
